refactoring home_body from home_screen
